backend/requirements.txt
Django>=5.0,<6
django-filter>=24.0,<26
djangorestframework>=3.15,<4
drf-spectacular>=0.27,<0.29
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
python-dotenv>=1.0,<2
mysqlclient>=2.2,<3
